requirements.txt
Django>=4.2,<5.0
djangorestframework
mysqlclient
djangorestframework-simplejwt
python-dotenv
